feat: divide tests suites
