+++ suite/src/__tests__/fast/model-correctness.test.ts
@@ -55,60 +55,4 @@ describe('Model Integration Test', () => {
     const document2 = await ModelInstanceDocument.load(ceramicNode2, document1.id)
     expect(document2.id).toEqual(document1.id)
   })
-
-  test.only('Model instance document changes are delivered via the datafeed SSE Api', async () => {
-    const modelInstanceDocumentMetadata = { model: modelId }
-    const Codec = JsonAsString.pipe(AggregationDocument)
-
-    const source1 = new EventSource(
-      new URL('/api/v0/feed/aggregation/documents', ComposeDbUrls[0]).toString(),
-    )
-    const source2 = new EventSource(
-      new URL('/api/v0/feed/aggregation/documents', ComposeDbUrls[1]).toString(),
-    )
-
-    const parseEventData = (eventData: any) => {
-      const decoded = decode(Codec, eventData)
-      return decoded.commitId.commit.toString()
-    }
-
-    const accumulator1 = new EventAccumulator(source1, parseEventData)
-    const accumulator2 = new EventAccumulator(source2, parseEventData)
-
-    try {
-      const expectedEvents = new Set()
-      const document1 = await ModelInstanceDocument.create(
-        ceramicNode1,
-        { myData: 40 },
-        modelInstanceDocumentMetadata,
-      )
-      expectedEvents.add(document1.tip.toString())
-
-      const document2 = await ModelInstanceDocument.create(
-        ceramicNode1,
-        { myData: 50 },
-        modelInstanceDocumentMetadata,
-      )
-      expectedEvents.add(document2.tip.toString())
-
-      const document3 = await ModelInstanceDocument.create(
-        ceramicNode1,
-        { myData: 60 },
-        modelInstanceDocumentMetadata,
-      )
-      expectedEvents.add(document3.tip.toString())
-
-      await document1.replace({ myData: 41 })
-      expectedEvents.add(document1.tip.toString())
-      await document2.replace({ myData: 51 })
-      expectedEvents.add(document2.tip.toString())
-      await document1.replace({ myData: 42 })
-      expectedEvents.add(document1.tip.toString())
-      await accumulator1.waitForEvents(expectedEvents, 1000 * 60)
-      await accumulator2.waitForEvents(expectedEvents, 1000 * 60)
-    } finally {
-      source1.close()
-      source2.close()
-    }
-  })
 })
